Refactor the Redis client initialization
